+++ ServerDevcommands/Commands/Teleport.cs
@@ -13,7 +13,7 @@ public class TeleportCommand
       if (index == 0) return ParameterInfo.PlayerNames;
       if (index == 1) return ParameterInfo.XZY("Target coordinates", subIndex);
       if (index == 2) return ParameterInfo.Create("Target rotation in degrees.");
-      if (index == 3) return ParameterInfo.Flag("fast");
+      if (index == 3) return ParameterInfo.Flag("fast", subIndex);
       return ParameterInfo.None;
     });
     Helper.Command("tp", "[player1,player2,...] [x,z,y/player] [rotY=0] [fast=false] - Teleports the player to coordinates or another player.", (args) =>
@@ -46,7 +46,7 @@ public class TeleportCommand
       var fastTeleport = args.Length > 4 && args[4].Equals("fast", StringComparison.OrdinalIgnoreCase);
       foreach (var player in players)
       {
-        ZRoutedRpc.instance.InvokeRoutedRPC(player.Character, "RPC_TeleportPlayer", [pos, rot, !fastTeleport]);
+        ZRoutedRpc.instance.InvokeRoutedRPC(player.SessionId, "RPC_TeleportPlayer", [pos, rot, !fastTeleport]);
       }
     });
   }