+++ publish/CHANGELOG.md
@@ -1,6 +1,9 @@
 - v1.68
+  - Fixes the `tp` command not working.
+  - Fixes not working with latest Comfy Gizmo.
   - Internal changes for new Infinity Hammer and Infinity Tools mods.
   - Removes the bind tagging system as obsolete.
+  - Removes setting "Disable parameter warnings" as obsolete.
 
 - v1.67
   - Adds a new command `search_component` for debugging purposes.